Change the return type of the PokemonDetailsService
Instead of returning the full pokemon details as they are returned
by the pokemon api we just return a subset of the data. This isolates
better other layers of the applications from changes in the remote
pokemon rest api. If it change returning different data the changes
will be isolated to the service